feature: Split variables and util definitions from main py file
