+++ .gitignore
@@ -1,1 +1,2 @@
 myenv
+py_scripts/__pycache__
