+++ py_scripts/variables.py
@@ -1,0 +1,7 @@
+#######
+# Microsoft Azure SQL Database
+#######
+DB_USERNAME = 'carlosarango97'
+DB_SERVER = 'codechallengeca.database.windows.net'
+DB_PASSWORD = 'Ca-97080504'
+DB_NAME = 'company'
